Collaborative Accelerated Failure Time Analysis (CAFTA)

Fits an accelerated failure time (AFT) model using collaborative analysis for survival data.

Usage

cafta(
  formula,
  data,
  dist = c("weibull", "loglogistic", "lognormal"),
  robust = FALSE,
  init,
  ...
)

Arguments

formula

A formula object specifying the model. The response must be a survival object created using the Surv function from the survival package.

data

A data frame containing the variables in the model.

dist

A character string specifying the distribution to be used. Choices are "weibull", "loglogistic", or "lognormal". Default is "weibull".

robust

Logical; if TRUE, computes robust (sandwich) covariance estimates. Default is FALSE.

init

Optional numeric vector of initial values for the parameters. If not supplied, defaults are used.

...

Additional arguments (currently unused).

Value

An object of class "cafta" containing:

theta

Estimated parameter vector.

hess

Estimated Hessian matrix.

cov

Estimated covariance matrix.

formula

The model formula.

dist

The distribution used.

robust

Logical indicating if robust covariance was used.

call

The matched call.

Details

This function fits an accelerated failure time model using maximum likelihood estimation. The supported distributions are Weibull, log-logistic, and log-normal. Robust covariance estimation is available via the robust argument.

Examples

formula <- Surv(time, status) ~ x1 + x2 + x31 + x42 + x43 + x44
fit <- cafta(formula, sim[sim$group == 1, ], dist = "weibull")
for (k in 2:6) {
  fit <- update(fit, sim[sim$group == k, ])
}
summary(fit)
#> Call:
#> update.cafta(object = fit, newdata = sim[sim$group == k, ])
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  0.34509    0.10260   3.363  0.00077 ***
#> x1           0.15768    0.03265   4.829 1.37e-06 ***
#> x2          -0.22310    0.03408  -6.546 5.92e-11 ***
#> x31          0.31520    0.07759   4.062 4.86e-05 ***
#> x42         -0.07377    0.12250  -0.602  0.54706    
#> x43         -0.13745    0.10652  -1.290  0.19691    
#> x44          0.17426    0.11507   1.514  0.12992    
#> log(scale)  -0.23404    0.03479  -6.727 1.74e-11 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1